Connector: Load custom interventions into model
Adjust the calculation of intervention dates, covering a domain of [0..365].
Add an import guard for data.wl in model.wl.
This allows the connector to import data.wl, mutate the global state, and then import model.wl without losing the changes to the global state.
Use this mechanism to add a custom scenario ID and distancing function.